refactor: Move Mandrill webhook from maillog to mail package
