+++ .env.example
@@ -130,7 +130,7 @@ SEND_MAILS_SUBJECT_PREFIX=Test
 #MANDRILL_API_KEY=
 
 # Allows to accept and verify Mandrill Webhook calls
-#MANDRILL_WEBHOOK_URL=http://localhost:5010/maillog/mandrill/webhook
+#MANDRILL_WEBHOOK_URL=http://localhost:5010/mail/mandrill/webhook
 #MANDRILL_WEBHOOK_KEY=xxx
 
 DEFAULT_MAIL_FROM_ADDRESS="kontakt@republik.ch"
